requirements.txt
discord.py
python-dotenv~=1.1.0
translationstring~=25.0.1
flask